New: reorganise classes into folders
