+++ templates/branded-emails/email-template.php
@@ -9,6 +9,8 @@
 
 namespace Eighteen73\Orbit;
 
+use Eighteen73\Orbit\Utilities\Templates;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
